requirements.txt
Django==5.0.4
djangorestframework==3.15.1
python-dotenv===1.0.1
djangorestframework-simplejwt===5.3.1
psycopg2-binary===2.9.9
django-filter===24.2
markdown===3.6
django-cors-headers===4.3.1
django-crontab===0.7.1